refactor: mv structs out of lib.rs
